forgot one change to long from double

+++ app/src/main/java/com/a0xffffffff/dinesum/UserTracker.java
@@ -32,7 +32,7 @@ public class UserTracker {
         return -1;
     }
 
-    public User updateUsersPoints(String userID, double points) {
+    public User updateUsersPoints(String userID, long points) {
         for (User u : mAllUsers) {
             if (u.getUserID().equals(userID)) {
                 u.setPoints(u.getPoints() + points);
